treewide: migrate from @asyncio.coroutine and yield from
Renames wpull.testing.async to wpull.testing._async, because async has
become a reserved term in 3.7.

The @asyncio.coroutine decorator was deprecated in 3.8 and removed in
3.10. Coroutines now use `async def fn` syntax.

The `yield from` expression in a coroutine was replaced by `await`
in 3.5.

The `asyncio.async` was an alias to `ensure_future`, deprecated in 3.4.4
and removed whenever.

And instead of `with (yield from lock)` locks now use the
`async with lock` construction.